Cleared compiler warnings and improved the cross-platform compatibility

+++ src/RangeCoder.cpp
@@ -78,7 +78,7 @@ int RangeCompress(const unsigned short * src, void * dest,int length, EncoderPai
 		prev = bit<<8;
 	};
 
-	return ending-(unsigned char *)dest;
+	return (int)(ending-(unsigned char *)dest);
 }
 
 int RangeDecompress( const void * source, unsigned short * dest,int length, DecoderPair * decoder_table, const int * const hash ){
